Add hint buttons in menus
