+++ client/src/ui/modules/military/Military.tsx
@@ -6,11 +6,16 @@ import { EntitiesArmyTable } from "@/ui/components/military/EntitiesArmyTable";
 import { useLocation } from "wouter";
 import { BattlesArmyTable } from "@/ui/components/military/BattlesArmyTable";
 import { Tabs } from "@/ui/elements/tab";
+import CircleButton from "@/ui/elements/CircleButton";
+import { HintModal } from "@/ui/components/hints/HintModal";
+import { BuildingThumbs } from "../navigation/LeftNavigationModule";
+import { useModal } from "@/hooks/store/useModal";
 
 export const Military = ({ entityId }: { entityId: bigint | undefined }) => {
   const [selectedTab, setSelectedTab] = useState(0);
 
   const { playerStructures } = useEntities();
+  const { toggleModal } = useModal();
 
   const [location, _] = useLocation();
 
@@ -43,7 +48,13 @@ export const Military = ({ entityId }: { entityId: bigint | undefined }) => {
   );
 
   return (
-    <div>
+    <div className="relative">
+      <CircleButton
+        className="absolute top-1 right-1"
+        onClick={() => toggleModal(<HintModal initialActiveSection={"Combat"} />)}
+        size={"sm"}
+        image={BuildingThumbs.question}
+      />
       {isMap ? (
         <Tabs
           selectedIndex={selectedTab}
